+++ examples/data-sources/meraki_appliance_single_lan/data-source.tf
@@ -1,0 +1,3 @@
+data "meraki_appliance_single_lan" "example" {
+  network_id = "L_123456"
+}