+++ examples/resources/meraki_appliance_single_lan/import.sh
@@ -1,0 +1,1 @@
+terraform import meraki_appliance_single_lan.example "<network_id>"