+++ CHANGELOG.md
@@ -22,6 +22,7 @@
 - Add `meraki_appliance_organization_security_intrusion` resource and data source
 - Add `meraki_appliance_security_malware` resource and data source
 - Add `meraki_appliance_settings` resource and data source
+- Add `meraki_appliance_single_lan` resource and data source
 
 ## 0.1.1